# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小黄成长日记",
        "title": "LarsWebv0",
        "contact": {
            "name": "Laurus",
            "email": "2312593392@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/ping": {
            "get": {
                "consumes": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "ping"
                ],
                "summary": "ping",
                "responses": {}
            }
        },
        "/user/login": {
            "post": {
                "consumes": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/x-json-stream"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "model.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "is_station": {
                    "type": "string"
                },
                "pass_word": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        }
    }
}